+++ src/components/NotificationBell.vue
@@ -1,8 +1,8 @@
-<template>
+<template functional>
   <li>
-    <a class="nav-item-iconic" href="javascript:void(0)" v-bind="$attrs" v-on="$listeners">
+    <a class="nav-item-iconic" href="javascript:void(0)" v-bind="data.attrs" v-on="listeners">
       <pf-icon name="fa-bell" title="Notifications"/>
-      <span class="badge" v-if="updates !== false"><span v-html="updates"></span></span>
+      <span class="badge" v-if="props.updates !== false"><span v-html="props.updates"></span></span>
     </a>
   </li>
 </template>